requirements.txt
# ==============================================================================
# PR Review Agent - Python Dependencies
# ==============================================================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# GitHub Integration
PyGithub==2.1.1
PyJWT==2.8.0
cryptography==42.0.0

# LLM Providers
openai==1.10.0
anthropic==0.18.0
# azure-openai==1.0.0  # Uncomment if using Azure

# Static Analysis Tools
flake8==7.0.0
pylint==3.0.3
bandit==1.7.6
radon==6.0.1
mccabe==0.7.0

# Code Quality
black==24.1.1
isort==5.13.2
mypy==1.8.0

# Diff and Code Parsing
unidiff==0.7.5
pygments==2.17.2

# AWS / Cloud Storage
boto3==1.34.34
botocore==1.34.34

# Data Processing
python-dotenv==1.0.1
pyyaml==6.0.1
toml==0.10.2

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-httpx==0.29.0

# Observability and Monitoring
structlog==24.1.0
python-json-logger==2.0.7
# sentry-sdk==1.40.0  # Uncomment for Sentry integration
# prometheus-client==0.19.0  # Uncomment for Prometheus metrics

# Security
python-multipart==0.0.6

# Database (Optional - for future use)
# sqlalchemy==2.0.25
# alembic==1.13.1
# asyncpg==0.29.0  # For PostgreSQL

# Caching (Optional)
# redis==5.0.1
# hiredis==2.3.2

# Utilities
click==8.1.7
python-dateutil==2.8.2
tenacity==8.2.3

# Type Hints
types-requests==2.31.0.20240125
types-PyYAML==6.0.12.12

# Development Tools
ipython==8.21.0
ipdb==0.13.13